created: list of similar movies and UI for similar movies list row refactor: API Error created: button like created: component in shared folder
