+++ Challenge-M2Y/Model/Movie.swift
@@ -10,9 +10,9 @@ import SwiftUI
 struct Movie: Decodable{
     var genres: [Genres]
     var release_date: String
-    var vote_count: Int // ammount of likes
-    var poster_path: String? // poster image
-    var backdrop_path: String? // image
+    var vote_count: Int 
+    var poster_path: String?
+    var backdrop_path: String?
     var popularity: Double
     var original_title: String
 }
